% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pill.R
\name{pill}
\alias{pill}
\title{Generate a shiny pill.}
\usage{
pill(id, label, type = "none", data_icon = NULL, class = NULL, ...)
}
\arguments{
\item{id}{\code{[character]}  All value must be unique.}

\item{label}{\code{[character]}  Must have an exact length of \code{length(id)}.  All value must be unique.}

\item{type}{\code{[any]}  Defaults to \code{'none'}}

\item{data_icon}{\code{[any]}  Defaults to \code{NULL}}

\item{class}{\code{[any]}  Defaults to \code{NULL}}
}
\value{
\code{pill}: \code{[HTML]}
}
\description{
Generate a shiny pill.
}
\examples{
 pill(id = 'id', label = 'ds', type = 'none', data_icon = NULL,
 class = NULL)
}
